feat: add contributors page and update routing for contributors and license sections
