+++ lib/src/common/routing/app_router.dart
@@ -6,6 +6,7 @@ import 'package:namma_wallet/src/features/bottom_navigation/presentation/namma_n
 import 'package:namma_wallet/src/features/calendar/presentation/calendar_view.dart';
 import 'package:namma_wallet/src/features/home/domain/generic_details_model.dart';
 import 'package:namma_wallet/src/features/home/presentation/home_view.dart';
+import 'package:namma_wallet/src/features/profile/presentation/contributors_page.dart';
 import 'package:namma_wallet/src/features/profile/presentation/db_viewer_page.dart';
 import 'package:namma_wallet/src/features/profile/presentation/license_view.dart';
 import 'package:namma_wallet/src/features/profile/presentation/profile_page.dart';
@@ -87,6 +88,11 @@ final router = GoRouter(
       path: AppRoute.license.path,
       name: AppRoute.license.name,
       builder: (context, state) => const LicenseView(),
+    ),
+    GoRoute(
+      path: AppRoute.contributors.path,
+      name: AppRoute.contributors.name,
+      builder: (context, state) => const ContributorsPage(),
     ),
   ],
 );
