+++ lib/src/features/profile/presentation/profile_page.dart
@@ -1,94 +1,12 @@
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:http/http.dart' as http;
 import 'package:namma_wallet/src/common/routing/app_routes.dart';
 import 'package:namma_wallet/src/common/theme/theme_provider.dart';
 import 'package:namma_wallet/src/common/widgets/custom_back_button.dart';
 import 'package:provider/provider.dart';
-import 'package:url_launcher/url_launcher.dart';
 
-// ----------------- Model -----------------
-class Contributor {
-  Contributor({
-    required this.name,
-    required this.avatarUrl,
-    required this.profileUrl,
-  });
-
-  factory Contributor.fromJson(Map<String, dynamic> json) {
-    return Contributor(
-      name: json['login'] as String,
-      avatarUrl: json['avatar_url'] as String,
-      profileUrl: json['html_url'] as String,
-    );
-  }
-  final String name;
-  final String avatarUrl;
-  final String profileUrl;
-}
-
-// ----------------- Profile Page -----------------
-class ProfilePage extends StatefulWidget {
+class ProfilePage extends StatelessWidget {
   const ProfilePage({super.key});
-
-  @override
-  State<ProfilePage> createState() => _ProfilePageState();
-}
-
-class _ProfilePageState extends State<ProfilePage> {
-  late Future<List<Contributor>> _contributorsFuture;
-
-  @override
-  void initState() {
-    super.initState();
-    _contributorsFuture = _fetchContributors();
-  }
-
-  Future<List<Contributor>> _fetchContributors() async {
-    final contributors = <Contributor>[];
-    var page = 1;
-    const perPage = 100;
-    const timeout = Duration(seconds: 10);
-
-    while (true) {
-      final uri = Uri.parse(
-        'https://api.github.com/repos/Namma-Flutter/namma_wallet/contributors',
-      ).replace(queryParameters: {
-        'per_page': perPage.toString(),
-        'page': page.toString(),
-      });
-
-      final response = await http.get(
-        uri,
-        headers: {
-          'User-Agent': 'namma_wallet',
-          'Accept': 'application/vnd.github.v3+json',
-        },
-      ).timeout(timeout);
-
-      if (response.statusCode == 200) {
-        final body = json.decode(response.body) as List<dynamic>;
-        if (body.isEmpty) {
-          break; // No more contributors
-        }
-        contributors.addAll(
-          body.map(
-            (json) => Contributor.fromJson(json as Map<String, dynamic>),
-          ),
-        );
-        page++;
-      } else {
-        throw Exception(
-          'Failed to load contributors: HTTP ${response.statusCode}\n'
-          'Response body: ${response.body}',
-        );
-      }
-    }
-
-    return contributors;
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -108,7 +26,40 @@ class _ProfilePageState extends State<ProfilePage> {
           const SizedBox(height: 24),
 
           // Contributors Section
-          ContributorsSectionWidget(contributorsFuture: _contributorsFuture),
+          Card(
+            elevation: 2,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(16),
+            ),
+            child: ListTile(
+              leading: const Icon(Icons.people_outline),
+              title: const Text('Contributors'),
+              subtitle: const Text('View project contributors'),
+              trailing: const Icon(Icons.chevron_right),
+              onTap: () {
+                context.pushNamed(AppRoute.contributors.name);
+              },
+            ),
+          ),
+
+          const SizedBox(height: 8),
+
+          // License Section
+          Card(
+            elevation: 2,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(16),
+            ),
+            child: ListTile(
+              leading: const Icon(Icons.article_outlined),
+              title: const Text('Licenses'),
+              subtitle: const Text('View open source licenses'),
+              trailing: const Icon(Icons.chevron_right),
+              onTap: () {
+                context.pushNamed(AppRoute.license.name);
+              },
+            ),
+          ),
 
           const SizedBox(height: 100), // Space for FAB
         ],
@@ -124,7 +75,6 @@ class _ProfilePageState extends State<ProfilePage> {
   }
 }
 
-// ----------------- Theme Section Widget -----------------
 class ThemeSectionWidget extends StatelessWidget {
   const ThemeSectionWidget({
     required this.themeProvider,
@@ -197,123 +147,6 @@ class ThemeSectionWidget extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-// ----------------- Contributors Section Widget -----------------
-class ContributorsSectionWidget extends StatelessWidget {
-  const ContributorsSectionWidget({
-    required this.contributorsFuture,
-    super.key,
-  });
-
-  final Future<List<Contributor>> contributorsFuture;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        const Padding(
-          padding: EdgeInsets.symmetric(horizontal: 4),
-          child: Row(
-            children: [
-              Icon(Icons.people_outline, size: 24),
-              SizedBox(width: 12),
-              Text(
-                'Contributors',
-                style: TextStyle(
-                  fontSize: 20,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ],
-          ),
-        ),
-        const SizedBox(height: 12),
-        FutureBuilder<List<Contributor>>(
-          future: contributorsFuture,
-          builder: (context, snapshot) {
-            if (snapshot.connectionState == ConnectionState.waiting) {
-              return const Card(
-                child: Padding(
-                  padding: EdgeInsets.all(32),
-                  child: Center(child: CircularProgressIndicator()),
-                ),
-              );
-            } else if (snapshot.hasError) {
-              return Card(
-                child: Padding(
-                  padding: const EdgeInsets.all(16),
-                  child: Column(
-                    children: [
-                      const Icon(Icons.error_outline,
-                          size: 48, color: Colors.red),
-                      const SizedBox(height: 8),
-                      Text('Error: ${snapshot.error}'),
-                    ],
-                  ),
-                ),
-              );
-            } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
-              return const Card(
-                child: Padding(
-                  padding: EdgeInsets.all(16),
-                  child: Center(child: Text('No contributors found.')),
-                ),
-              );
-            }
-
-            final contributors = snapshot.data!;
-            return Column(
-              children: contributors
-                  .map((contributor) => Card(
-                        margin: const EdgeInsets.only(bottom: 8),
-                        elevation: 1,
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(12),
-                        ),
-                        child: ListTile(
-                          leading: CircleAvatar(
-                            backgroundImage:
-                                NetworkImage(contributor.avatarUrl),
-                            radius: 24,
-                          ),
-                          title: Text(
-                            contributor.name,
-                            style: const TextStyle(fontWeight: FontWeight.w600),
-                          ),
-                          subtitle: Text(
-                            contributor.profileUrl,
-                            style: const TextStyle(fontSize: 12),
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                          trailing: const Icon(Icons.open_in_new, size: 20),
-                          onTap: () async {
-                            final url = Uri.parse(contributor.profileUrl);
-                            try {
-                              await launchUrl(url);
-                            } on Exception catch (_) {
-                              if (context.mounted) {
-                                ScaffoldMessenger.of(context).showSnackBar(
-                                  SnackBar(
-                                    content: Text(
-                                      'Could not open '
-                                      '${contributor.profileUrl}',
-                                    ),
-                                  ),
-                                );
-                              }
-                            }
-                          },
-                        ),
-                      ))
-                  .toList(),
-            );
-          },
-        ),
-      ],
     );
   }
 }
